add support IOS platfrom (First Step)

+++ nexign-task/src/main/java/io/github/artsok/tasks/repository/IndividualRepository.java
@@ -5,5 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IndividualRepository extends JpaRepository<Individual, Long> {
+public interface IndividualRepository extends JpaRepository<Individual, Integer> {
+
 }
